+++ packages/unmock-core/src/service/dsl/constants.ts
@@ -1,0 +1,3 @@
+const SCHEMA_PREFIX = "x-unmock-";
+export const SCHEMA_TIMES = `${SCHEMA_PREFIX}times`;
+export const UNMOCK_TYPE = "unmock";
